requirements.txt
anthropic>=0.52.0
chromadb>=0.4.0
openai>=1.82.0
python-dotenv>=1.1.0
